move ordinal to models
